HBX-1684: Move class 'org.hibernate.tool.hbm2x.visitor.DefaultValueVisitor' to 'org.hibernate.tool.hbm2x.visitor.DefaultValueVisitor'
