+++ orm/src/main/java/org/hibernate/tool/internal/export/common/EntityNameFromValueVisitor.java
@@ -1,4 +1,4 @@
-package org.hibernate.tool.hbm2x.visitor;
+package org.hibernate.tool.internal.export.common;
 
 import org.hibernate.mapping.Array;
 import org.hibernate.mapping.Bag;
@@ -14,6 +14,7 @@ import org.hibernate.mapping.PrimitiveArray;
 import org.hibernate.mapping.Set;
 import org.hibernate.mapping.SimpleValue;
 import org.hibernate.mapping.ToOne;
+import org.hibernate.tool.hbm2x.visitor.DefaultValueVisitor;
 
 public class EntityNameFromValueVisitor extends DefaultValueVisitor {
