+++ orm/src/main/java/org/hibernate/tool/internal/export/lint/BadCachingDetector.java
@@ -4,7 +4,7 @@ import org.hibernate.mapping.Collection;
 import org.hibernate.mapping.PersistentClass;
 import org.hibernate.mapping.Property;
 import org.hibernate.mapping.Value;
-import org.hibernate.tool.hbm2x.visitor.EntityNameFromValueVisitor;
+import org.hibernate.tool.internal.export.common.EntityNameFromValueVisitor;
 
 public class BadCachingDetector extends EntityModelDetector {
